Separate globals file responsabilities
