+++ js/dom/lib/schema.js
@@ -1,0 +1,18 @@
+import { getCurrentSchemaVersion } from "../../globals.js";
+
+export function schemaWithLinks() {
+  const currentSchemaVersion = getCurrentSchemaVersion();
+
+  if (typeof currentSchemaVersion === "undefined") {
+    return false;
+  }
+
+  if (currentSchemaVersion === "old") {
+    return false;
+  }
+  if (currentSchemaVersion == 1) {
+    return false;
+  }
+
+  return true;
+}
